[FIX] working expiration_panel
