+++ database_block/static/src/js/expiration_panel.js
@@ -1,72 +1,18 @@
 /** @odoo-module **/
 
-import { ExpirationPanel } from "@web_enterprise/static/src/webclient/home_menu";
-import { session } from "@web/session";
+import { ExpirationPanel } from "@web_enterprise/webclient/home_menu/expiration_panel";
 import { patch } from "@web/core/utils/patch";
-// const { patch } = require("web.utils");
 
-export class DatabaseBlock extends ExpirationPanel {
-    constructor() {
-        super(...arguments);
-        this.enterprise = useService("enterprise");
+const { hooks } = owl;
+const { useState } = hooks;
 
-        // Implementar un patch:
-        // https://www.odoo.com/documentation/15.0/es/developer/reference/frontend/patching_code.html#description
-
-        patch(ExpirationPanel.prototype, "patching_my_component", {
-            setup() {
-                this._super(...arguments);
-                console.log("patch patch paatch ...");
-                console.log(session);
-            },
+patch(ExpirationPanel.prototype, "setup mounted", {
+    mounted() {
+        this._super(...arguments);
+        this.state = useState({
+            ...this.state,
+            database_block_message: this.enterprise.blockMessage
         });
-        // if (!this.enterprise.warning) {
-        //     this.state = { display: false };
-        //     return;
-        // }
-
-        // this.cookie = useService("cookie");
-        // this.rpc = useService("rpc");
-        // this.orm = useService("orm");
-
-        // let expirationDate;
-        // if (this.enterprise.expirationDate) {
-        //     expirationDate = deserializeDateTime(this.enterprise.expirationDate);
-        // } else {
-        //     // If no date found, assume 1 month and hope for the best
-        //     expirationDate = DateTime.utc().plus({ days: 30 });
-        // }
-        // const diffDays = this._computeDiffDays(expirationDate);
-
-        // const hideCookie = this.cookie.current.oe_instance_hide_panel;
-
-        // let alertType = "info";
-        // if (diffDays <= 5) {
-        //     alertType = "danger";
-        // } else if (diffDays <= 7) {
-        //     alertType = "warning";
-        // }
-
-        // this.state = useState({
-        //     display: (diffDays <= 10 && !hideCookie) || diffDays <= 0,
-        //     alertType,
-        //     buttonText: "Register",
-        //     diffDays,
-        //     message: "register",
-        //     expirationDate,
-        // });
-
-        // this.inputRef = useRef("input");
-
-        // // Expiration reason e.g. 'trial','renewal','upsell',...
-        // this.expirationReason = this.enterprise.expirationReason;
-        // this.notYetRegistered = ["trial", "demo", false].includes(
-        //     this.expirationReason
-        // );
-
-        // this.isMailInstalled = this.enterprise.isMailInstalled;
-
-        // // Type of logged-in accounts addressed by message
-        // this.warning = this.enterprise.warning;
-    }
-}
+        console.log(this.state);
+    },
+});
